Minor - Further Dialogue Updates
Fixed issue where the dialogue will no longer appear off to the left or
right of the screen.

Need to refactor the script code to only pass through the text box and
the object speaking, as it is too difficult to implement how to position
the text box currently. The template script generate text based on x,y
location, but the box varies in height which causes boxes to overlap and
I can only fix the y position in a draw task which I assume makes it
resource intensive.