The one with dialogue refactoring
I changed the scr_text so that you only need to pass the text, speed of
the dialogue, and object which is speaking. the script will handle
placement of the text box based on object height and how close the text
box is to the edge of the screen. I also looked into changing the text
draw so it would not draw parts of words before hitting a new line, but
it will be way easier to just manually add the breaks where needed.

I think we also need a new font, as the one I have in there now will not
properly render apostrophes.